refactor(routes): parseFeedback を services/feedback.ts に分離
ルートには Repository 呼び出し・エラーハンドリング・レスポンス整形のみ残し、
DB の JSON カラム変換ロジックを Service 層に移動。list は Promise.all で並列化。
